backend/requirements.txt
aioboto3==10.2.0
aiobotocore==2.4.1
aiohttp==3.8.6
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.18.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
black==24.3.0
boto3==1.24.59
botocore==1.27.59
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.88.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
jmespath==1.0.1
kafka-python==2.0.2
motor==3.1.2
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
Pillow==9.3.0
platformdirs==4.2.0
pluggy==1.4.0
pydantic==1.10.14
pymongo==4.3.3
pypng==0.20220715.0
pytest==7.2.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
qrcode==7.4.2
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.1
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.10.0
urllib3==1.26.18
uvicorn==0.20.0
wrapt==1.16.0
yarl==1.9.4
